app/requirements.txt
python-telegram-bot==20.1
python-dotenv==1.0.0
aws-cdk-lib==2.77.0
awscli==1.27.131
boto3==1.26.131
mongopersistence==0.3.0